moved CASA topic to config/topics, added labels layer on top of other CASA layers
